fix search by custName
