+++ packages/kokoas-client/src/pages/projProspectSearchV2/filter/Filter.tsx
@@ -1,19 +1,18 @@
-import { Button, Stack, TextField } from '@mui/material';
+import { Button, Stack } from '@mui/material';
 import { useTypedFormContext } from '../hooks/useTypedHooks';
 import { DevTool } from '@hookform/devtools';
 import { Keyword } from './Keyword';
 import { SearchButton } from './SearchButton';
 import { FilterButton } from './FilterButton';
-import { useStartSearch } from '../hooks/useStartSearch';
 import { useNavigate } from 'react-router-dom';
 import { useSnackBar } from 'kokoas-client/src/hooks';
 import { CocoAGSearch } from './CocoAGSearch';
+import { ControlledTextField } from './fields/ControlledTextField';
 
 
 export const Filter = () => {
-  const { control, register } = useTypedFormContext();
+  const { control } = useTypedFormContext();
   const navigate = useNavigate();
-  const handleStartSearch = useStartSearch();
   const { setSnackState } = useSnackBar();
 
 
@@ -23,17 +22,12 @@ export const Filter = () => {
       direction={'row'}
     >
       <Keyword />
-      <TextField 
-        size="small" 
-        label="お客様名"
+      <ControlledTextField 
+        label='お客様名'
+        name='custName'
         placeholder='氏名・シメイ'
-        onKeyDown={(e) => {
-          if (e.key === 'Enter') {
-            handleStartSearch();
-          }
-        }}
-        {...register('custName')}
       />
+      
       <CocoAGSearch />
       <SearchButton />
       <FilterButton />
